pyproject.toml
[project]
name = "fastapi-strands-hexagonal-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "strands-agents[otel]>=1.4.0",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
]

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
ignore = ["E402"]
